% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{register_task_definition}
\alias{register_task_definition}
\title{Title}
\usage{
register_task_definition(region, family, image, execution_role, task_role,
  cpu, memory, log_group = family, log_stream_prefix = family)
}
\arguments{
\item{log_stream_prefix}{}
}
\description{
Title
}
