% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{run_task}
\alias{run_task}
\title{Title}
\usage{
run_task(region, family, cluster, subnets, security_groups = list(),
  assign_public_ip = F, revision = NULL, count = 1,
  environment = list(), command = c())
}
\arguments{
\item{environment}{}
}
\description{
Title
}
